Move rollcall to interfaces
